// src/app/models/Worker.ts
export class Worker{
    id: number;
    name: string;
    surname: string;
    patronymic: string;
    phone: string;
    dateOfBirth: string;
    city: string;
    street: string;
    house: string;
    familyPosition: string;
    countChildren: number;
    email: string;
    seriesPass: number;
    numberPass: number;
    issuedByWhom: string;
    dateOfIssue: string;
    divisionCode: string;
    numberSnils: number;
    numberInn: number;
    gender: string;
    militaryTitle: string;
    shelfLife: string;
    stockCategory: number;
    profile: string;
    vus: string;
    nameKomis: string;
    photo: string;
    descriptionWorker: string;
    dismiss: boolean;


    namePost: String;

}